+++ src/server/devices/groups/Fenster.ts
@@ -1,7 +1,7 @@
 import { HmIpGriff } from '../hmIPDevices';
 import { LogDebugType, ShutterService, TimeCallbackService, Utils } from '../../services';
 import { FensterPosition } from '../models';
-import { LogLevel, TimeOfDay } from '../../../models';
+import { LogLevel, TimeOfDay, WindowSettings } from '../../../models';
 import { iShutter, iVibrationSensor } from '../baseDeviceInterfaces';
 import { BaseGroup } from './base-group';
 import { GroupType } from './group-type';
@@ -11,6 +11,7 @@ import { ZigbeeMagnetContact } from '../zigbee';
 
 export class Fenster extends BaseGroup {
   public desiredPosition: number = 0;
+  public settings: WindowSettings = new WindowSettings();
 
   public constructor(
     roomName: string,